Mix pose factor as popup when clicking thumbnail
When you click a pose thumbnail a popup for the mix factor is shown.
When you change the factor, you can immediately see the result.

Next up: try to do it with a modal operator, so you don't even need the
popup. Just click, drag and release for changing the mix factor or just
click to apply the pose 100%.