Press SHIFT to start mixing the pose
Now you have to press SHIFT to mix the pose instead of applying it. Just
applying is a faster on complex rigs and I don't know a good way to
check if the thumbnail is just clicked or it is a 'click and drag'
operation. This way the user has to distinctly choose what he wants.