Rename mouse_prev_x to mouse_rev_x + calculate in different place
The variable indicates the reference point for the mouse movements, and
not necessarily the 'previous' X-coordinate of the mouse.

Also this is better calculated in the modal() function, as execute() should
just use the properties (so self.factor) and not look at mouse position.
This makes it possible to simply call the operator in a non-modal way and
pass a factor.