Show drop-down selector for pose libraries for current character only.
The character name is assumed to be the object name until the first
dash, space, etc. (i.e. /^[a-z0-9_]+/). This naming convention is quite
specific to the Blender Animation Studio, so maybe at some point the
regexp and the action prefix for pose libraries should be part of the
add-on preferences.