Fix: don't save the state of the 'Apply Flipped' checkbox
This fixes "Apply Flipped" being checked but the thumbnails being shown
normally. This happens when "Apply Flipped" was checked before saving.

This option shouldn't be saved, because the on_flipped_update() function
dumbly flips the pixels of the images. It assumes that the loaded images
are consistent with this `flipped` property, whereas Blender will start
up with the images loaded as on-disk (i.e. as if flipped=False,
regardless of the `flipped` property).