Fixed use of Rig Prefix and class loading, added update method for All Poses.
Updated the draw panels as well as the rename_for_character operator to respect the optional rig prefix, and updated draw code to be more consistent between both views. Fixed an issue where registration would sometimes due to the random order in which classes are loaded. This was due to nested property groups, wherein the property group  registration depends on  being registered first which was not guaranteed before. Finally, added an update method to the bool property All Poses so that thumbnails are automatically redrawn if toggles, as it may be confusing to the user in the properties panel, and in the 3D view panel there furthermore isn't a way to refresh the thumbnails before at all (rendering the checkbox pointless before, as you could not refresh the thumbnails from that panel).